Added mouse detection.
In InputHandler I have added a way for the mouse's position to be detected, in a std::pair<int,int>. Also, I have added left and right click detection (no scroll wheel). Finally I added a way to see if the click was just a click, and not a hold over multiple frames.