Created Trigger
I've created a trigger. This will react when a sprite comes within a configured range of it. It will then call 3 separate callback functions.

onEnter - when the sprite enters the radius of the trigger, this function will be called.

onOverlapping - when the sprite is inside the radius, every update frame will call this function.

onExit - when the sprite exits the radius, this callback function will be executed.

Option to enable and disable the trigger.