GameObject Removal Runtime
GameObjects can now be removed at runtime without seg faults. If you wish to delete the object (i.e allocated on heap) then use the deleteAtFrameEnd(GameObject *address) signature. This way that the object is deleted at the end of the frame, rather than during the updates. This means that I don't have to constantly check for nullptr.